+++ Editor/SettingsAnalysis/PlayerSettingsAnalyzer.cs
@@ -17,6 +17,8 @@ namespace Unity.ProjectAuditor.Editor.SettingsAnalysis
         internal const string PAS0002 = nameof(PAS0002);
         internal const string PAS0029 = nameof(PAS0029);
         internal const string PAS0033 = nameof(PAS0033);
+        internal const string PAS1004 = nameof(PAS1004);
+        internal const string PAS1005 = nameof(PAS1005);
 
         static readonly Descriptor k_AccelerometerDescriptor = new Descriptor(
             PAS0002,
@@ -48,11 +50,42 @@ namespace Unity.ProjectAuditor.Editor.SettingsAnalysis
             })
         };
 
+        static readonly Descriptor k_IL2CPPCompilerConfigurationMasterDescriptor = new Descriptor(
+            PAS1004,
+            "Player: IL2CPP Compiler Configuration",
+            new[] { Area.BuildTime },
+            "<b>C++ Compiler Configuration</b> is set to <b>Master</b>. This mode is intended for shipping builds and will significantly increase build times.",
+            "To reduce build times, change <b>Project Settings ➔ Configuration ➔ C++ Compiler Configuration</b> to <b>Release</b>.")
+        {
+            fixer = (issue => {
+                SetIL2CPPConfigurationToRelease();
+            }),
+
+            messageFormat = "Player: C++ Compiler Configuration is set to 'Master'"
+        };
+
+        static readonly Descriptor k_IL2CPPCompilerConfigurationDebugDescriptor = new Descriptor(
+            PAS1005,
+            "Player: IL2CPP Compiler Configuration",
+            new[] { Area.CPU },
+            "<b>C++ Compiler Configuration</b> is set to <b>Debug</b>. This mode is intended for debugging and might have an impact on runtime performance.",
+            "To improve runtime performance, change <b>Project Settings ➔ Configuration ➔ C++ Compiler Configuration</b> to <b>Release</b>.")
+        {
+            fixer = (issue =>
+            {
+                SetIL2CPPConfigurationToRelease();
+            }),
+
+            messageFormat = "Player: C++ Compiler Configuration is set to 'Debug'"
+        };
+
         public void Initialize(ProjectAuditorModule module)
         {
             module.RegisterDescriptor(k_AccelerometerDescriptor);
             module.RegisterDescriptor(k_SplashScreenDescriptor);
             module.RegisterDescriptor(k_SpeakerModeDescriptor);
+            module.RegisterDescriptor(k_IL2CPPCompilerConfigurationMasterDescriptor);
+            module.RegisterDescriptor(k_IL2CPPCompilerConfigurationDebugDescriptor);
         }
 
         public IEnumerable<ProjectIssue> Analyze(ProjectAuditorParams projectAuditorParams)
@@ -70,6 +103,16 @@ namespace Unity.ProjectAuditor.Editor.SettingsAnalysis
             if (!IsSpeakerModeMono())
             {
                 yield return ProjectIssue.Create(IssueCategory.ProjectSetting, k_SpeakerModeDescriptor)
+                    .WithLocation("Project/Player");
+            }
+            if (CheckIL2CPPCompilerConfiguration(Il2CppCompilerConfiguration.Master, projectAuditorParams))
+            {
+                yield return ProjectIssue.Create(IssueCategory.ProjectSetting, k_IL2CPPCompilerConfigurationMasterDescriptor)
+                    .WithLocation("Project/Player");
+            }
+            if (CheckIL2CPPCompilerConfiguration(Il2CppCompilerConfiguration.Debug, projectAuditorParams))
+            {
+                yield return ProjectIssue.Create(IssueCategory.ProjectSetting, k_IL2CPPCompilerConfigurationDebugDescriptor)
                     .WithLocation("Project/Player");
             }
         }
@@ -106,6 +149,24 @@ namespace Unity.ProjectAuditor.Editor.SettingsAnalysis
             };
 
             AudioSettings.Reset(audioConfiguration);
+        }
+
+        internal static bool CheckIL2CPPCompilerConfiguration(Il2CppCompilerConfiguration compilerConfiguration, ProjectAuditorParams projectAuditorParams)
+        {
+            var buildTargetGroup = BuildPipeline.GetBuildTargetGroup(projectAuditorParams.platform);
+            if (PlayerSettings.GetScriptingBackend(buildTargetGroup) !=
+                ScriptingImplementation.IL2CPP)
+            {
+                return false;
+            }
+
+            return PlayerSettings.GetIl2CppCompilerConfiguration(buildTargetGroup) ==
+                compilerConfiguration;
+        }
+
+        internal static void SetIL2CPPConfigurationToRelease()
+        {
+            PlayerSettings.SetIl2CppCompilerConfiguration(EditorUserBuildSettings.selectedBuildTargetGroup, Il2CppCompilerConfiguration.Release);
         }
     }
 }